% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalancedSampling-package.R
\docType{package}
\name{BalancedSampling}
\alias{BalancedSampling}
\alias{BalancedSampling-package}
\title{BalancedSampling: Balanced and Spatially Balanced Sampling}
\description{
Select balanced and spatially balanced probability samples in multi-dimensional
spaces with any prescribed inclusion probabilities.
It contains fast (C++ via Rcpp) implementations of the included sampling methods.

You can access the project website at
\url{https://envisim.se}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.envisim.se/}
  \item \url{https://github.com/envisim/BalancedSampling/}
}

}
\author{
Anton Grafström \email{anton.grafstrom@gmail.com}, Jonathan Lisic, Wilmer Prentius.
}
\keyword{internal}
