% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpm.R
\name{rpm}
\alias{rpm}
\title{Random Pivotal Method}
\usage{
rpm(prob, eps = 1e-12)
}
\arguments{
\item{prob}{A vector of length N with inclusion probabilities.}

\item{eps}{A small value used to determine when an updated probability is
close enough to 0.0 or 1.0.}
}
\value{
A vector of selected indices in 1,2,...,N.
}
\description{
Selects samples with prescribed inclusion probabilities from a finite population.
This design has high entropy.
}
\details{
If \code{prob} sum to an integer n, a fixed sized sample (n) will be produced.
In each of the (at most) N steps, two undecided units are selected at random
to compete.
}
\examples{
\dontrun{
set.seed(12345);
N = 1000;
n = 100;
prob = rep(n/N, N);
x = matrix(runif(N * 2), ncol = 2);
s = rpm(prob);
plot(x[, 1], x[, 2]);
points(x[s, 1], x[s, 2], pch = 19);

set.seed(12345);
prob = c(0.2, 0.25, 0.35, 0.4, 0.5, 0.5, 0.55, 0.65, 0.7, 0.9);
N = length(prob);
ep = rep(0L, N);
r = 10000L;
for (i in seq_len(r)) {
  s = rpm(prob);
  ep[s] = ep[s] + 1L;
}
print(ep / r);
}

}
\seealso{
Other sampling: 
\code{\link{cubestratified}()},
\code{\link{cube}()},
\code{\link{hlpm2}()},
\code{\link{lcubestratified}()},
\code{\link{lcube}()},
\code{\link{lpm1search}()},
\code{\link{lpm1}()},
\code{\link{lpm2}()},
\code{\link{scps}()},
\code{\link{spm}()}
}
\concept{sampling}
