% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb.R
\name{sb}
\alias{sb}
\title{Spatial balance}
\usage{
sb(prob, x, sample, measure = "voronoi", type = "kdtree2", bucketSize = 10)
}
\arguments{
\item{prob}{A vector of length N with inclusion probabilities, or an integer > 1.
If an integer n, then the probabilities will be assumed to be equal n / N.}

\item{x}{An N by p matrix of (standardized) auxiliary variables.}

\item{sample}{A vector of sample indices.}

\item{measure}{The type of balance measure to use.
Must be one of \code{"voronoi"}.}

\item{type}{The method used in finding nearest neighbours.
Must be one of \code{"kdtree0"}, \code{"kdtree1"}, \code{"kdtree2"}, and
\code{"notree"}.}

\item{bucketSize}{The maximum size of the terminal nodes in the k-d-trees.}
}
\description{
Calculates the spatial balance of a sample.
}
\examples{
\dontrun{
set.seed(12345);
N = 500;
n = 70;
prob = rep(n / N, N);
x = matrix(runif(N * 2), ncol = 2);
s = lpm2(prob, x);
b = sb(prob, x, s);
}
}
\references{
Grafström, A., Lundström, N.L.P. & Schelin, L. (2012).
Spatially balanced sampling through the Pivotal method.
Biometrics 68(2), 514-520.

Friedman, J. H., Bentley, J. L., & Finkel, R. A. (1977).
An algorithm for finding best matches in logarithmic expected time.
ACM Transactions on Mathematical Software (TOMS), 3(3), 209-226.

Maneewongvatana, S., & Mount, D. M. (1999, December).
It’s okay to be skinny, if your friends are fat.
In Center for geometric computing 4th annual workshop on computational geometry (Vol. 2, pp. 1-8).

Lisic, J. J., & Cruze, N. B. (2016, June).
Local pivotal methods for large surveys.
In Proceedings of the Fifth International Conference on Establishment Surveys.
}
